% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addnoise_spd.R
\name{addNoise_spd}
\alias{addNoise_spd}
\title{addNoise_spd}
\usage{
addNoise_spd(A, SNR = 1)
}
\arguments{
\item{A}{An SPD matrix which will be noisified.}

\item{SNR}{Signal to Noise ratio to be used in generating the noise. SNR must be larger 0.}
}
\description{
Adds random noise to an SPD matrix, given a signal to noise ratio. Signal and Noise are measured as distance on the SPD manifold. If A is the SPD matrix, N is the noise matrix, and I is the identity matrix, then Signal=dist(I,A), Noise=dist(A,N)), and SNR = Signal/Noise.
}
\examples{
set.seed(623766)
A = randspd_FAST(n=5)
SNR=.25
addNoise_spd(A,SNR=SNR)
}
