% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addnoise_spd.R
\name{addrelnoise_spd}
\alias{addrelnoise_spd}
\title{add noise relative to A where maxerr is percentage error (in relation to A)}
\usage{
addrelnoise_spd(A, maxerr)
}
\description{
add noise relative to A where maxerr is percentage error (in relation to A)
}
