% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mglm_algo.R
\name{aug3}
\alias{aug3}
\title{aug3}
\usage{
aug3(M)
}
\arguments{
\item{M}{M is a matrix which should be augmented to a flat array.}
}
\description{
Augment a n x m matrix to a n x m x 1 array, inverse of cont3.
}
