% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechet_variance_spd.R
\name{calc_Rsqr_spd}
\alias{calc_Rsqr_spd}
\title{calc_Rsqr_spd}
\usage{
calc_Rsqr_spd(Y, Yhat, pKarcher = NULL, maxiter = 200)
}
\arguments{
\item{Y}{The observed SPD matrices (an nxnxN array of N SPD matrices)}

\item{Yhat}{The estimated SPD matrices (an nxnxN array of N SPD matrices)}

\item{pKarcher}{The Karcher mean of the observed SPD matrices, or NULL if the Karcher mean is to be calculated}

\item{maxiter}{If pKarcher is NULL, maxiter is the maximum number of iterations for calculating the Karcher mean.}
}
\description{
Calculates R Squared on the SPD manifold.
}
