% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mglm_algo.R
\name{cont3}
\alias{cont3}
\title{cont3}
\usage{
cont3(M)
}
\arguments{
\item{M}{M is a flat array which should be contracted down to a matrix.}
}
\description{
Contract an n x m x 1 array to an n x m matrix, inverse of aug3.
}
