% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_M_spd.R
\name{dist_M_euc}
\alias{dist_M_euc}
\title{dis_M_euc}
\usage{
dist_M_euc(X, Y)
}
\arguments{
\item{X}{a matrix}

\item{Y}{a matrix}
}
\description{
Calculates the distance in euclidean space between two matrices X and Y. \eqn{d = trace(X'Y)^{0.5}}
}
