% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_M_spd.R
\name{dist_M_spd}
\alias{dist_M_spd}
\title{dis_M_spd}
\usage{
dist_M_spd(X, Y)
}
\arguments{
\item{X}{an SPD matrix}

\item{Y}{an SPD matrix}
}
\description{
Calculates the distance on the SPD manifold between two SPD matrices X and Y. \eqn{<Log(X,Y)|Log(X,Y)>_{X}^{0.5}}
}
