% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expmap_spd.R
\name{expmap_spd}
\alias{expmap_spd}
\title{expmap_spd}
\usage{
expmap_spd(P, X)
}
\arguments{
\item{P}{A point on the SPD manifold (a symmetric positive definite matrix).}

\item{X}{A symmetric matrix (a matrix in the tangent space of the SPD manifold).}
}
\description{
Exponential Map projects a tangent vector X onto the SPD manifold at the point P. Analagous to P+X in euclidean space.
}
