% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feval_spd.R
\name{feval_spd}
\alias{feval_spd}
\title{feval_spd}
\usage{
feval_spd(p, V, X, Y)
}
\arguments{
\item{p}{p is the base point.}

\item{V}{V is a set of symmetric matrices.}

\item{X}{X is a set of column vectors.}

\item{Y}{Y is a set of SPD matrices.}
}
\description{
Calculates the sum of squared geodesic errors (which mglm_spd() tries to minimize).
}
