% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechet_variance_spd.R
\name{frechet_variance_spd}
\alias{frechet_variance_spd}
\title{Frechet Variance}
\usage{
frechet_variance_spd(Y, pKarcher = NULL, maxiter = 200)
}
\arguments{
\item{Y}{dxdxn array of dxd spd matrices to find the frechet variance of}

\item{pKarcher}{if not NULL, the karcher mean from which to center the variance. If NULL, the karcher mean will be calculated.}

\item{maxiter}{if pKarcher is NULL, maxiter is the maximum number of iterations for calculating the karcher mean.}
}
\description{
Frechet Variance
}
