% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isspd.R
\name{isspd}
\alias{isspd}
\title{isspd}
\usage{
isspd(mx, c = .Machine$double.eps)
}
\arguments{
\item{mx}{Matrix to check.}

\item{c}{The smallest eigenvalue threshold (eigenvalue larger than c implies non-zero eigenvalue, and SPD matrices have all non-zero eigenvalues).}
}
\description{
Checks whether a matrix is symmetric positive definite (SPD) or not.
}
