% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isspd_mxstack.R
\name{isspd_mxstack}
\alias{isspd_mxstack}
\title{isspd_mxstack}
\usage{
isspd_mxstack(Y)
}
\arguments{
\item{Y}{A dxdxN array of N matrices to check.}
}
\description{
Checks an array of matrices to ensure each matrix is symmetric positive definite.
}
