% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isspd.R
\name{issym}
\alias{issym}
\title{issym}
\usage{
issym(mx)
}
\arguments{
\item{mx}{Matrix to check.}
}
\description{
Checks whether a matrix is symmetric or not. Note that this function takes advantage of the package Rfast if it is installed.
}
