% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karcher_mean_spd.R
\name{karcher_mean_spd}
\alias{karcher_mean_spd}
\title{karcher_mean_spd}
\usage{
karcher_mean_spd(X, W = NULL, niter)
}
\arguments{
\item{X}{X is a dxdxN array of dxd SPD matrices, for which to calculate the Karcher mean.}

\item{W}{W is a vector of N weights to apply to the SPD matrices (if W == NULL then equal weights are applied).}

\item{niter}{niter is the maximum number of iterations in calculating the Karcher mean.}
}
\description{
Calculates the Karcher mean with weights W on the SPD manifold.
}
