% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logmap_spd.R
\name{logmap_spd}
\alias{logmap_spd}
\title{logmap_spd}
\usage{
logmap_spd(P, X)
}
\arguments{
\item{P}{P is a point on the SPD manifold (a positive definite symmetric matrix).}

\item{X}{X is an SPD matrix to be mapped to the tangent space at P.}
}
\description{
Logarithmic Map takes X from the SPD manifold and maps it to the tangent space of symmetric matrices at the point P on the SPD manifold. Analagous to X-P in euclidean space.
}
