% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_TpM_spd.R
\name{norm_TpM_spd}
\alias{norm_TpM_spd}
\title{norm_TpM_spd}
\usage{
norm_TpM_spd(p, v)
}
\arguments{
\item{p}{p is a point on the SPD manifold.}

\item{v}{v is a tangent vector for which to calculate the norm on the SPD manifold.}
}
\description{
Calculates the norm of the symmetric matrix v on the SPD manifold at the point p.
}
