% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paralleltranslateAtoB_spd.R
\name{paralleltranslateAtoB_spd}
\alias{paralleltranslateAtoB_spd}
\title{paralleltranslateAtoB_spd}
\usage{
paralleltranslateAtoB_spd(a, b, w)
}
\arguments{
\item{a}{The starting point on the SPD manifold (a symmetric positive definite matrix).}

\item{b}{The ending point on the SPD manifold (a symmetric positive definite matrix).}

\item{w}{The set of tangent vectors (symmetrix matrices) which should be moved via parallel transport from a to b.}
}
\description{
Performs a parallel transport of the symmetrix matrices w, from point a to point b on the SPD manifold.
}
