% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_spd.R
\name{prediction_spd}
\alias{prediction_spd}
\title{prediction_spd}
\usage{
prediction_spd(p, V, X)
}
\arguments{
\item{p}{p is a base point (SPD maxtrix).}

\item{V}{V is a set of tangent vectors (dxdxdim(X) symmetric matrices).}

\item{X}{X is a set of covariates, dimX x N column vectors.}
}
\description{
Given a base point p, a set of tangent vectors V, and covariates X, calculates Y. Used to transform model estimates p_hat, V_hat, X_hat into model predictions Y_hat.
}
