% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_M_spd.R
\name{proj_M_spd}
\alias{proj_M_spd}
\title{proj_M_spd}
\usage{
proj_M_spd(X, c = .Machine$double.eps)
}
\arguments{
\item{X}{A matrix to transform into an SPD matrix.}

\item{c}{c is the eigenvalues threshold (if any eigenvalue is less than c, then the matrix is not SPD, since an SPD matrix has all non-zero eigenvalues).}
}
\description{
Projects a matrix X on to the SPD manifold.
}
