% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_TpM_spd.R
\name{proj_TpM_spd}
\alias{proj_TpM_spd}
\title{proj_TpM_spd}
\usage{
proj_TpM_spd(V)
}
\arguments{
\item{V}{A dxdxN array of dxd matrices to project onto the space of symmetric matrices.}
}
\description{
Projects a set of matrices V onto the tangent space of the SPD manifold (transforms each matrix into a symmetric matrix).
}
