% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randspd.R
\name{randspd_FAST}
\alias{randspd_FAST}
\title{randspd_FAST}
\usage{
randspd_FAST(n, maxDist = 3, showDist = F)
}
\arguments{
\item{n}{dimensions of nxn SPD matrix}

\item{maxDist}{maximum distance of SPD matrix from I_nxn}

\item{showDist}{if T, will print the distance of random SPD maxtrix from I_nxn}
}
\description{
Generate a random Symmetric Positive Definite matrix of dimension n, such that the maximum distance on the manifold is less than maxDist. Makes use of the sparsebnUtils::random.spd function from the library sparsebnUtils.
}
