% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwrite_spd.R
\name{readSPD}
\alias{readSPD}
\title{readSPD}
\usage{
readSPD(filename = "mat.spd")
}
\arguments{
\item{filename}{name of file containing SPD matrix data.}
}
\description{
read a file containing SPD matrices written using writeSPD().
}
