% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mglm_algo.R
\name{sizeR}
\alias{sizeR}
\title{sizeR}
\usage{
sizeR(M, d = NULL)
}
\arguments{
\item{M}{M is a matrix or array object.}

\item{d}{d is the dimension for which a size is requested (d==NULL returns a vector of all dimsension sizes).}
}
\description{
Calculates size of each dimension d of a matrix or array M, returns 1 if d is NA
}
