% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_spd.R
\name{SSE_spd}
\alias{SSE_spd}
\title{SSE_spd}
\usage{
SSE_spd(Y, Yhat)
}
\arguments{
\item{Y}{dxdxN array of observed dxd SPD matrices.}

\item{Yhat}{dxdxN array of estimated dxd SPD matrices.}
}
\description{
Calculate the error sum of squares on the SPD manifold.
}
