% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_spd.R
\name{SSR_spd}
\alias{SSR_spd}
\title{SSR_spd}
\usage{
SSR_spd(Yhat, Ybar)
}
\arguments{
\item{Yhat}{dxdxN array of estimated dxd SPD matrices.}

\item{Ybar}{Ybar is the Karcher mean of the SPD matrices Y.}
}
\description{
Calculate the regression sum of squares on the SPD manifold.
}
