% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_spd.R
\name{SST_spd}
\alias{SST_spd}
\title{SST_spd}
\usage{
SST_spd(Y, Ybar = NULL, niter = 200)
}
\arguments{
\item{Y}{dxdxN array of observed dxd SPD matrices.}

\item{Ybar}{Ybar is the Karcher mean of the SPD matrices Y.}

\item{niter}{number of iterations used to calculate Karcher mean (if Ybar==NULL).}
}
\description{
Calculate the total sum of squares on the SPD manifold.
}
