% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwrite_spd.R
\name{writeSPD}
\alias{writeSPD}
\title{writeSPD}
\usage{
writeSPD(Y, ID = NULL, filename = "mat.spd")
}
\arguments{
\item{Y}{a dxdxN array of SPD matrices}

\item{ID}{a vector of ID's associated with the SPD matrices (defaults to integers 1:N).}
}
\description{
write dxdxN array of SPD matrices to disk as a data.frame.
}
