% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addnoise_spd.R
\name{addSNR_spd}
\alias{addSNR_spd}
\title{add noise relative to A where SNR is the signal (A) to noise ratio}
\usage{
addSNR_spd(A, SNR)
}
\description{
add noise relative to A where SNR is the signal (A) to noise ratio
}
