% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechet_variance_spd.R
\name{frechet_variance_spd}
\alias{frechet_variance_spd}
\title{frechet_variance_spd}
\usage{
frechet_variance_spd(Y, pKarcher = NULL, maxiter = 200)
}
\arguments{
\item{Y}{dxdxn array of dxd SPD matrices for which to find the Frechet Variance.}

\item{pKarcher}{The Karcher mean from which to center the variance. If NULL, the Karcher mean will be calculated.}

\item{maxiter}{if pKarcher is NULL, maxiter is the maximum number of iterations for calculating the Karcher mean.}
}
\description{
Calculates the Frechet Variance for a set of SPD matrices.
}
